Move `Camera`, `Viewport`, and `GraphicsOptions` actually into `all-is-cubes-render`.
This reduces the amount of code in the `all-is-cubes` crate, and reduces
the number of items whose public paths are not in the crate they are
defined in.

The prior reason for keeping them in `all-is-cubes` was for the
raytracer, but that has since been moved too.
