+++ all-is-cubes-render/src/camera/mod.rs
@@ -3,37 +3,29 @@
 
 use all_is_cubes::euclid::Size2D;
 
-#[cfg(doc)]
-use all_is_cubes::space::Space;
-
 // -------------------------------------------------------------------------------------------------
 
-/// Defines a perspective view in/of the world.
-///
-/// A [`Camera`] has the following independently controllable properties.
-///
-/// * A [`ViewTransform`], which specifies the viewpoint (eye position) and
-///   direction.
-/// * A [`Viewport`], whose aspect ratio is used in the projection matrix.
-/// * A [`GraphicsOptions`], whose `fov_y` is used in the projection matrix.
-/// * A “measured exposure” (luminance scale factor) value which should be provided by examining
-///   the world, and which is carried through to [the output](Camera::exposure).
-///
-/// From this information, it derives [view](Camera::view_matrix) and
-/// [projection](Camera::projection_matrix) matrices.
-///
-/// It does not specify *what* [`Space`] is to be viewed; more broadly, it is
-/// plain data structure that does some calculations, and knows nothing outside of that.
-/// See [`StandardCameras`] for going further.
-pub use all_is_cubes::camera::Camera;
-
 // As a workaround for <https://github.com/rust-lang/rust/issues/127445>,
-// we list some items explicitly even though they should be redundant with
-// the glob re-exports.
-pub use all_is_cubes::camera::{GraphicsOptions, ToneMappingOperator, ViewTransform, Viewport, *};
+// we list all items explicitly and avoid cross-crate glob re-exports.
+pub use all_is_cubes::camera::{Eye, ImagePixel, ImageSize, ViewTransform, eye_for_look_at};
+
+mod camera_struct;
+pub use camera_struct::{Camera, Ndc, NdcPoint2, NdcPoint3};
+
+mod graphics_options;
+pub use graphics_options::{
+    AntialiasingOption, ExposureOption, FogOption, GraphicsOptions, LightingOption, RenderMethod,
+    ToneMappingOperator, TransparencyOption,
+};
 
 mod stdcam;
 pub use stdcam::*;
+
+mod viewport;
+pub use viewport::{NominalPixel, Viewport};
+
+#[cfg(test)]
+mod tests;
 
 // -------------------------------------------------------------------------------------------------
